+++ src/pages/IssueToken/hooks/useIssueTokenForm.ts
@@ -1,9 +1,3 @@
-import {
-  Address,
-  TokenManagementTransactionsFactory,
-  TransactionsFactoryConfig
-} from '@multiversx/sdk-core/out';
-
 import BigNumber from 'bignumber.js';
 import { useFormik } from 'formik';
 import { object, string } from 'yup';
@@ -12,7 +6,10 @@ import {
   useGetAccount,
   useGetNetworkConfig,
   maxDecimals,
-  stringIsFloat
+  stringIsFloat,
+  Address,
+  TokenManagementTransactionsFactory,
+  TransactionsFactoryConfig
 } from 'lib';
 
 import { DECIMALS } from 'localConstants';
@@ -70,30 +67,25 @@ export const useIssueTokenForm = () => {
         )
     }),
     onSubmit: async (values) => {
-      try {
-        const transaction = factory.createTransactionForIssuingFungible({
-          sender: new Address(address),
-          tokenName: values.tokenName,
-          tokenTicker: values.tokenTicker.toUpperCase(),
-          initialSupply: BigInt(
-            new BigNumber(values.mintedValue)
-              .pow(10, values.numDecimals)
-              .toNumber()
-          ),
-          numDecimals: BigInt(values.numDecimals),
-          canFreeze: true,
-          canWipe: true,
-          canPause: true,
-          canChangeOwner: true,
-          canUpgrade: true,
-          canAddSpecialRoles: true
-        });
+      const transaction = factory.createTransactionForIssuingFungible({
+        sender: new Address(address),
+        tokenName: values.tokenName,
+        tokenTicker: values.tokenTicker.toUpperCase(),
+        initialSupply: BigInt(
+          new BigNumber(values.mintedValue)
+            .pow(10, values.numDecimals)
+            .toNumber()
+        ),
+        numDecimals: BigInt(values.numDecimals),
+        canFreeze: true,
+        canWipe: true,
+        canPause: true,
+        canChangeOwner: true,
+        canUpgrade: true,
+        canAddSpecialRoles: true
+      });
 
-        await sendTransactions([transaction]);
-      } catch (err) {
-        //setErrors({ amount: err.message });
-      }
-
+      await sendTransactions([transaction]);
       formik.resetForm();
     }
   });