+++ src/pages/index.tsx
@@ -3,6 +3,8 @@ export * from './SovereignTransfer';
 export * from './Disclaimer';
 export * from './Home';
 export * from './Hook';
+export * from './IssueCollection';
+export * from './IssueNft';
 export * from './IssueToken';
 export * from './Logout';
 export * from './PageNotFound';