+++ src/redux/endpoints/nfts.endpoint.ts
@@ -3,9 +3,11 @@ import {
   ACCOUNTS_ENDPOINT,
   API_CACHE_DURATION_SECONDS,
   MAX_API_SIZE,
-  NFTS_ENDPOINT
+  NFTS_ENDPOINT,
+  ROLES_ENDPOINT,
+  COLLECTIONS_ENDPOINT
 } from 'localConstants';
-import { GetNftsType, PartialNftType } from 'types';
+import { GetNftsType, PartialNftType, CollectionType } from 'types';
 import { RootApi } from '../rootApi';
 
 const nftsEndpoints = RootApi.injectEndpoints({
@@ -19,8 +21,22 @@ const nftsEndpoints = RootApi.injectEndpoints({
         method: 'GET',
         params: { size: MAX_API_SIZE, ...props }
       })
+    }),
+    getCollections: builder.query<CollectionType[], string>({
+      keepUnusedDataFor: API_CACHE_DURATION_SECONDS,
+      query: (address) => ({
+        baseURL: getBaseURL(),
+        url: `/${ACCOUNTS_ENDPOINT}/${address}/${ROLES_ENDPOINT}/${COLLECTIONS_ENDPOINT}`,
+        method: 'GET',
+        params: { creator: address }
+      })
     })
   })
 });
 
-export const { useGetNftsQuery, useLazyGetNftsQuery } = nftsEndpoints;
+export const {
+  useGetNftsQuery,
+  useGetCollectionsQuery,
+  useLazyGetNftsQuery,
+  useLazyGetCollectionsQuery
+} = nftsEndpoints;